+++ plotly/src/scatter.rs
@@ -11,7 +11,7 @@ use serde::Serialize;
 
 #[cfg(feature = "plotly_ndarray")]
 use crate::ndarray::ArrayTraces;
-use crate::private::{copy_iterable_to_vec, NumOrString, TruthyEnum};
+use crate::private::{copy_iterable_to_vec, NumOrString};
 #[cfg(feature = "plotly_ndarray")]
 use ndarray::{Array, Ix1, Ix2};
 
@@ -25,7 +25,7 @@ where
     #[serde(skip_serializing_if = "Option::is_none")]
     name: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    visible: Option<TruthyEnum<Visible>>,
+    visible: Option<Visible>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "showlegend")]
     show_legend: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "legendgroup")]
@@ -276,7 +276,7 @@ where
     /// Determines whether or not this trace is visible. If `Visible::LegendOnly`, the trace is not
     /// drawn, but can appear as a legend item (provided that the legend itself is visible).
     pub fn visible(mut self, visible: Visible) -> Box<Self> {
-        self.visible = Some(TruthyEnum { e: visible });
+        self.visible = Some(visible);
         Box::new(self)
     }